NOTE: interp%ipow and interp%%ipow2 are declared as real(kind=8)    instead of integer.  dTWdx2 and dTWdxmdxn are implemented.
   CONCERNS: Tay*d2Wdx2 and Tay*d2Wdxmdxn need to be doube checked

+++ Potentials/MSI/interp_mod.F90
@@ -6,7 +6,7 @@ module interpolation
   !---------------------------------------------
   type interp_params
     integer :: ipart,neigh_update,nneigh,ngroup
-    integer :: ipow,lowp, ipow2
+    real(kind=8) :: ipow,lowp, ipow2
     integer :: ndata,nforc
     real(kind=8) :: wtol,outer,vmax, vmin, sigma
     real(kind=8) :: avrads
